+++ CHANGELOG.md
@@ -3,7 +3,10 @@ Initial release.
 
 ## v0.1.1
 Bug fixes:
- * make sure argv[0] is a absolute path in icppcli driver;
- * fix c source file error compiling arguments;
+ * icpp: make sure argv[0] is a absolute path in icppcli driver;
+ * icpp: fix c source file error compiling arguments;
 
 Improvements:
+ * icpp: add icpp module header directory to clang header search path;
+ * imod: add executable binary packing support;
+ * imod: add apple framework packing support;